+++ ContentPatcher/docs/release-notes.md
@@ -13,7 +13,8 @@ When releasing a format change, don't forget to update the smapi.io/json schema!
 * Added new token:
   * [`PathPart`](author-tokens-guide.md#PathPart): get part of a file/asset path (e.g. for patches with multiple `Target` or `FromFile` values).
 * Improved tokens:
-  * The default tokens which take a `PlayerType` argument now accept a specific player ID too, like `{{PlayerName: 3864039824286870457}}`.
+  * Tokens with a player type argument now recognize player IDs too (like `{{PlayerName: 3864039824286870457}}`).
+  * You can now specify an optional player type/ID with more tokens (specifically `IsMainPlayer`, `IsOutdoors`, `LocationContext`, `LocationName`, `LocationUniqueName`, `PlayerGender`, `PlayerName`, and `Spouse`).
 * Improved context updates:
   * The default tokens are now available immediately after the save file is read, before the game even starts loading it.
   * The `Hearts` and `Relationships` tokens are now available before the save is fully loaded (like the other built-in tokens).